% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{compile.Flist}
\alias{compile.Flist}
\title{this function makes an index for each feature that indicates whether it has a genbank, refseq, both or neither annotation file. Values: 0 = neither; 1 = GenBank only; 2 = RefSeq only; 3 = both.}
\usage{
compile.Flist(refdir, path1, path2)
}
\arguments{
\item{refdir}{table of NCBI prokaryote assemblies}

\item{path1}{path to GenBank folder}

\item{path2}{path to RefSeq folder}
}
\description{
This is an internal subroutine for get.genes
}
\examples{
compile.Flist()
}
