% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{download.Feature.Tables}
\alias{download.Feature.Tables}
\title{this function unzips the annotation features (gene annotations) for each assembly in refseq and genbank}
\usage{
download.Feature.Tables(refdir, outpath, binPATH = "/usr/local/bin/")
}
\arguments{
\item{refdir}{path to table of all complete prokaryote genomes from NCBI, with ftp links. Can be downloaded from https://www.ncbi.nlm.nih.gov/genome/browse#!/prokaryotes/. For example: "~/Documents/prokaryotes.csv"}

\item{outpath}{= output directory}

\item{binPath}{= path to bin where wget is compiled.}
}
\description{
this function unzips the annotation features (gene annotations) for each assembly in refseq and genbank
}
\examples{
download.Feature.Tables()
}
