% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{get.genes}
\alias{get.genes}
\title{This function is a wrapper that implements the unzip, compile.Flist and compileFunctionTable functions}
\usage{
get.genes(refdir, directory, genes)
}
\arguments{
\item{refdir}{table of NCBI prokaryote assemblies}

\item{directory}{directory that contains the downloaded feature annotation tables}

\item{gene}{list of gene names, e.g. c("nifH", "nifD")}
}
\description{
This function is a wrapper that implements the unzip, compile.Flist and compileFunctionTable functions
}
\examples{
get.genes()
}
