% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{match.gene}
\alias{match.gene}
\title{subroutine for compile.Functiontable}
\usage{
match.gene(x, type, file1, file2 = NULL)
}
\arguments{
\item{x}{gene symbol to search for}

\item{type}{whether assemblies for one or both databases exist. 1 = only one (RefSeq or GenBank); 2=both exist}
}
\description{
Used to construct the index column to indicate whether GenBank or RefSeq annotation files exist
}
\examples{
match.gene()
}
