% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{RankAbundanceBurnoulliPlot}
\alias{RankAbundanceBurnoulliPlot}
\title{create taxonomy vs rank-abundance curves}
\usage{
RankAbundanceBurnoulliPlot(ps)
}
\arguments{
\item{ps}{phyloseq object}
}
\description{
create taxonomy vs rank-abundance curves
}
\examples{
RankAbundanceBurnoulliPlot()
}
