% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{subset.Phyloseq}
\alias{subset.Phyloseq}
\title{Use gene annotations to subset phyloseq object}
\usage{
\method{subset}{Phyloseq}(ps, reftab, genes, either = T, level = "Species", cutoff = 1)
}
\arguments{
\item{ps}{phyloseq object}

\item{reftab}{reference table built from get.genes()}

\item{genes}{genes to use in the subset}

\item{either}{logical, if T then returns taxa that have at least one gene in the list; if F then returns only taxa that have both}

\item{level}{one of "Genus" or "Species" to select from dataset}

\item{cutoff}{threshold - proportion of assemblies that meet genes criteria for a taxon to keep in the reference taxon. 1 represents all assemblies meet the gene critera}
}
\description{
This function is not finished
}
\examples{
subset.phyloseq()
}
