% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{unzip}
\alias{unzip}
\title{this function unzips the annotation features (gene annotations) for each assembly in refseq and genbank}
\usage{
unzip(x, a)
}
\arguments{
\item{x}{list of files}

\item{a}{1 or 2 for GenBank or RefSeq}
}
\description{
This is an internal subroutine for get.genes
}
\examples{
unzip()
}
