% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Functions.R
\name{compile.Functiontable}
\alias{compile.Functiontable}
\title{Compiles summary table of which assemblies have each gene.}
\usage{
compile.Functiontable(x, y)
}
\arguments{
\item{x}{element from refdir$Assembly. This is implemented within ldply}
}
\description{
Output table has columns for Genus, Species, Strain, Assembly ID, and each of the genes of interest
}
\examples{
compile.Functiontable()
}
\keyword{all}
\keyword{be}
\keyword{metadata}
\keyword{reference}
\keyword{table}
\keyword{the}
\keyword{will}
\keyword{with}
\keyword{y}
